Code refactoring performed to pass code review
