+++ core/src/main/java/com/exadel/aem/toolkit/core/injectors/utils/InstantiationUtil.java
@@ -21,10 +21,8 @@ import java.util.stream.StreamSupport;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.resource.Resource;
-import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.resource.ValueMap;
-import org.apache.sling.api.wrappers.DeepReadValueMapDecorator;
-import org.apache.sling.api.wrappers.ModifiableValueMapDecorator;
+import org.apache.sling.api.wrappers.ValueMapDecorator;
 import org.apache.sling.jcr.resource.api.JcrResourceConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -56,9 +54,9 @@ public class InstantiationUtil {
         try {
             return type.getConstructor().newInstance();
         } catch (InstantiationException
-            | IllegalAccessException
-            | InvocationTargetException
-            | NoSuchMethodException ex) {
+                 | IllegalAccessException
+                 | InvocationTargetException
+                 | NoSuchMethodException ex) {
             LOG.error("Could not initialize object {}", type.getName(), ex);
         }
         return null;
@@ -92,11 +90,14 @@ public class InstantiationUtil {
                 child.getResourceType(),
                 child.getValueMap()))
             .collect(Collectors.toList());
+        String resourceType = values.getOrDefault(JcrResourceConstants.SLING_RESOURCE_TYPE_PROPERTY, current.getResourceType()).toString();
+        ValueMap valueMap = new ValueMapDecorator(values);
+        ValueMap deepReadValueMap = new SubNameDeepReadValueMapDecorator(current, valueMap, prefix, postfix);
         return new ValueMapResource(
             current.getResourceResolver(),
             current.getPath(),
-            values.getOrDefault(JcrResourceConstants.SLING_RESOURCE_TYPE_PROPERTY, current.getResourceType()).toString(),
-            new SubNameDeepReadValueMapDecorator(current, new ModifiableValueMapDecorator(values), prefix, postfix),
+            resourceType,
+            deepReadValueMap,
             children);
     }
 
@@ -140,9 +141,9 @@ public class InstantiationUtil {
 
     /**
      * Removes the given prefix and/or postfix from the provided resource path
-     * @param path String value representing the path
-     * @param prefix   String value representing an optional prefix
-     * @param postfix  String value representing an optional postfix
+     * @param path    String value representing the path
+     * @param prefix  String value representing an optional prefix
+     * @param postfix String value representing an optional postfix
      * @return String value
      */
     private static String clearPrefixOrPostfixFromPath(String path, String prefix, String postfix) {
@@ -158,64 +159,5 @@ public class InstantiationUtil {
             name = StringUtils.removeEnd(name, postfix);
         }
         return parentPath + CoreConstants.SEPARATOR_SLASH + name;
-    }
-}
-
-class SubNameDeepReadValueMapDecorator extends DeepReadValueMapDecorator {
-    private final String prefix;
-
-    private final String postfix;
-
-    private final ValueMap base;
-
-    private final ResourceResolver resolver;
-
-    private final String pathPrefix;
-
-    public SubNameDeepReadValueMapDecorator(Resource resource, ValueMap base, String prefix, String postfix) {
-        super(resource, base);
-        this.resolver = resource.getResourceResolver();
-        this.pathPrefix = resource.getPath() + CoreConstants.SEPARATOR_SLASH;
-        this.base = base;
-        this.prefix = prefix;
-        this.postfix = postfix;
-    }
-
-    private ValueMap getValueMap(final String name) {
-        final int position = name.lastIndexOf(CoreConstants.SEPARATOR_SLASH);
-        if ( position == -1 ) {
-            return this.base;
-        }
-        final Resource resource = this.resolver.getResource(this.pathPrefix + this.getNameWithPrefixAndPostfix(name.substring(0, position)));
-        if ( resource != null ) {
-            final ValueMap valueMap = resource.adaptTo(ValueMap.class);
-            if ( valueMap != null ) {
-                return valueMap;
-            }
-        }
-        return ValueMap.EMPTY;
-    }
-
-    @Override
-    public <T> T get(String name, Class<T> type) {
-        return this.getValueMap(name).get(this.getPropertyName(name), type);
-    }
-
-    private String getNameWithPrefixAndPostfix(String name) {
-        if (StringUtils.isNotEmpty(prefix)) {
-            name = prefix + name;
-        }
-        if (StringUtils.isNotEmpty(postfix)) {
-            name = name + postfix;
-        }
-        return name;
-    }
-
-    private String getPropertyName(final String name) {
-        final int pos = name.lastIndexOf(CoreConstants.SEPARATOR_SLASH);
-        if ( pos == -1 ) {
-            return name;
-        }
-        return name.substring(pos + 1);
     }
 }
